Extract bottom sheet into separate view (No user-visible changes)
