+++ Elsewhen/ContentView.swift
@@ -32,52 +32,11 @@ let logger = Logger(subsystem: Bundle.main.bundleIdentifier!, category: "times")
 struct ContentView: View {
     
     @State private var selectedDate = Date()
-    @State private var selectedFormatStyle: DateFormat = Self.dateFormats[0]
+    @State private var selectedFormatStyle: DateFormat = dateFormats[0]
     @State private var selectedTimeZone: String = TimeZone.current.identifier
-    @State private var showCopied: Bool = false
     @State private var showLocalTimeInstead: Bool = false
-    #if os(iOS)
-    @State private var selectionFeedbackGenerator: UISelectionFeedbackGenerator? = nil
-    @State private var notificationFeedbackGenerator: UINotificationFeedbackGenerator? = nil
-    #endif
-    
-    static private let dateFormats: [DateFormat] = [
-        DateFormat(icon: "calendar.badge.clock", name: "Full", code: .f),
-        DateFormat(icon: "calendar.badge.plus", name: "Full with Day", code: .F),
-        DateFormat(icon: "calendar", name: "Date only", code: .D),
-        DateFormat(icon: "clock", name: "Time only", code: .t),
-        DateFormat(icon: "clock.fill", name: "Time with seconds", code: .T),
-        DateFormat(icon: "clock.arrow.2.circlepath", name: "Relative", code: .R),
-    ]
-    
-    private func format(date: Date, in timezone: TimeZone) -> String {
-        let dateFormatter = DateFormatter()
-        switch selectedFormatStyle.code {
-        case .f:
-            dateFormatter.dateStyle = .long
-            dateFormatter.timeStyle = .short
-        case .F:
-            dateFormatter.dateStyle = .full
-            dateFormatter.timeStyle = .short
-        case .D:
-            dateFormatter.dateStyle = .long
-        case .t:
-            dateFormatter.dateStyle = .none
-            dateFormatter.timeStyle = .short
-        case .T:
-            dateFormatter.dateStyle = .none
-            dateFormatter.timeStyle = .medium
-        case .R:
-            let relativeFormatter = RelativeDateTimeFormatter()
-            let convertedNow = convert(date: Date(), from: TimeZone.current, to: timezone)
-            return relativeFormatter.localizedString(for: date, relativeTo: convertedNow)
-        }
-        return dateFormatter.string(from: date)
-    }
     
     private var discordFormat: String {
-        
-        
         var timeIntervalSince1970 = Int(selectedDate.timeIntervalSince1970)
         
         if let tz = TimeZone(identifier: selectedTimeZone) {
@@ -87,10 +46,6 @@ struct ContentView: View {
         }
         
         return "<t:\(timeIntervalSince1970):\(selectedFormatStyle.code.rawValue)>"
-    }
-    
-    private func formatTimeZoneName(_ zone: String) -> String {
-        zone.replacingOccurrences(of: "_", with: " ")
     }
     
     var body: some View {
@@ -125,7 +80,7 @@ struct ContentView: View {
                         .frame(minWidth: 0, maxWidth: 390)
                         
                         HStack(spacing: 5) {
-                            ForEach(Self.dateFormats, id: \.self) { formatStyle in
+                            ForEach(dateFormats, id: \.self) { formatStyle in
                                 Button(action: {
                                     self.selectedFormatStyle = formatStyle
                                 }) {
@@ -156,118 +111,7 @@ struct ContentView: View {
                     
                 }
                 
-                VStack {
-                    VStack {
-                        
-                        let selectedTimezoneIdentifier = TimeZone(identifier: selectedTimeZone)!
-                        
-                        if !showLocalTimeInstead {
-                            
-                            Text(format(date: convertSelectedDate(from: selectedTimezoneIdentifier, to: selectedTimezoneIdentifier), in: selectedTimezoneIdentifier))
-                                .multilineTextAlignment(.center)
-                                .font(.headline)
-                                .foregroundColor(.primary)
-                            
-                        } else {
-                            
-                            Text(format(date: convertSelectedDate(from: selectedTimezoneIdentifier, to: TimeZone.current), in: TimeZone.current))
-                                .multilineTextAlignment(.center)
-                                .font(.headline)
-                                .foregroundColor(.primary)
-                            
-                        }
-                        
-                        if selectedTimezoneIdentifier != TimeZone.current {
-                            
-                            Text(showLocalTimeInstead ? "for you (\(formatTimeZoneName(TimeZone.current.identifier)))" : "in \(formatTimeZoneName(selectedTimeZone))")
-                                .foregroundColor(.secondary)
-                            
-                        }
-                        
-                    }
-                    .padding(5)
-                    .contextMenu {
-                        ForEach(Self.dateFormats, id: \.self) { formatStyle in
-                            Button(action: {
-                                self.selectedFormatStyle = formatStyle
-                            }) {
-                                Label(formatStyle.name, systemImage: formatStyle.icon)
-                            }
-                        }
-                    }
-                    .onTapGesture {
-                        #if os(iOS)
-                        selectionFeedbackGenerator = UISelectionFeedbackGenerator()
-                        #endif
-                        withAnimation {
-                            showLocalTimeInstead = true
-                            #if os(iOS)
-                            selectionFeedbackGenerator?.selectionChanged()
-                            #endif
-                        }
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 3) {
-                            withAnimation {
-                                showLocalTimeInstead = false
-                            }
-                            #if os(iOS)
-                            selectionFeedbackGenerator = nil
-                            #endif
-                        }
-                    }
-                    
-                    Group {
-                        
-                        DiscordFormattedDate(text: discordFormat)
-                        
-                        Button(action: {
-                            #if os(iOS)
-                            notificationFeedbackGenerator = UINotificationFeedbackGenerator()
-                            notificationFeedbackGenerator?.prepare()
-                            #endif
-                            UIPasteboard.general.setValue(self.discordFormat,
-                                                          forPasteboardType: kUTTypePlainText as String)
-                            withAnimation {
-                                showCopied = true
-                                #if os(iOS)
-                                notificationFeedbackGenerator?.notificationOccurred(.success)
-                                #endif
-                            }
-                            DispatchQueue.main.asyncAfter(deadline: .now() + 1) {
-                                withAnimation {
-                                    showCopied = false
-                                }
-                                #if os(iOS)
-                                notificationFeedbackGenerator = nil
-                                #endif
-                            }
-                        }) {
-                            Text(showCopied ? "Copied ✓" : "Copy Discord Code")
-                                .font(.headline)
-                                .foregroundColor(.white)
-                        }
-                        .padding(.horizontal)
-                        .padding(.vertical, 5)
-                        .background(
-                            RoundedRectangle(cornerRadius: 15, style: .continuous)
-                                .fill(Color.accentColor)
-                        )
-                        .padding(.bottom, 8)
-                        
-                        Text("Date and time representative of components only; may not match exact Discord formatting.")
-                            .multilineTextAlignment(.center)
-                            .font(.caption2)
-                            .foregroundColor(.secondary)
-                        
-                    }
-                    .padding(.horizontal)
-                    
-                }
-                .padding(.vertical)
-                .frame(minWidth: 0, maxWidth: .infinity)
-                .background(
-                    Color(UIColor.secondarySystemBackground)
-                        .shadow(radius: 10)
-                )
+                ResultSheet(selectedDate: selectedDate, selectedTimeZone: selectedTimeZone, discordFormat: discordFormat, showLocalTimeInstead: $showLocalTimeInstead, selectedFormatStyle: $selectedFormatStyle)
                 
             }
             .edgesIgnoringSafeArea([.bottom, .horizontal])
@@ -289,9 +133,4 @@ struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         ContentView()
     }
-}
-
-func convert(date: Date, from initialTimezone: TimeZone, to targetTimezone: TimeZone) -> Date {
-    let offset = TimeInterval(targetTimezone.secondsFromGMT(for: date) - initialTimezone.secondsFromGMT(for: date))
-    return date.addingTimeInterval(offset)
 }
